Cleanup tcms/core/utils/
