+++ src/tcms/testplans/helpers/email.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 from django.conf import settings
 
-from tcms.core.utils.mailto import mailto
+from tcms.core.mailto import mailto
 
 
 def email_plan_update(plan):
